backend/workflows/requirements.txt
anthropic
openai
ddgs
requests
beautifulsoup4
python-dotenv
rich
pdfplumber
PyMuPDF
APScheduler
pytz
playwright
websocket-client
